+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Not yet released
 * Fix dev UI support for Quarkus 2
 * Fix support for non-public `@PostLoad` methods in entity view annotation processor
 * Fix lookup of constantified attributes in functional dependency analysis
+* Retain configurations added through `MetadataContributor` in Quarkus integration
 
 ### Backwards-incompatible changes